+++ defaults/controllers/_default/root/admin/servers/edit.js
@@ -1,0 +1,15 @@
+// --------------------
+// Servers resource controller
+// edit action
+// --------------------
+
+// exports
+
+// action definition
+exports = module.exports = {
+	// functions
+	
+	process: function() {
+		return this.route.actions.new.process.call(this);
+	}
+};