+++ defaults/controllers/_default/root/api/_default/all.js
@@ -1,0 +1,13 @@
+// --------------------
+// API namespace controller
+// Default action
+// --------------------
+
+// exports
+
+exports = module.exports = {
+	access: function() {
+		// allow access if admin user
+		return !!this.user.permissions.API;
+	}
+};