+++ defaults/controllers/_default/root/api/_index.js
@@ -1,0 +1,13 @@
+// --------------------
+// API namespace controller
+// Route definition
+// --------------------
+
+// exports
+
+// route definition
+exports = module.exports = {
+	// route params
+	type: 'namespace',
+	titleRoute: 'API'
+};