+++ defaults/controllers/_default/root/api/index.js
@@ -1,0 +1,17 @@
+// --------------------
+// API namespace controller
+// index action
+// --------------------
+
+// imports
+var defaultAction = require('./_default/all');
+
+// exports
+
+// action definition
+exports = module.exports = {
+	// functions
+	
+	//xxx should not need to be defined here but inheritance not working at present
+	access: defaultAction.access
+};