+++ changelog.md
@@ -151,3 +151,4 @@
 * Remove admin controller to execute commands on database
 * `authentication.makeHash()` method
 * Added Server model + controllers
+* HTTP authentication for servers