+++ changelog.md
@@ -153,3 +153,4 @@
 * Added Server model + controllers
 * HTTP authentication for servers
 * Permissions and roles for access through API
+* API for database actions